fix clear color

+++ src/gpu/vlk/vlk_swapchain.c
@@ -196,10 +196,10 @@ void _vlk_swapchain__begin_frame(_vlk_swapchain_t* swap)
 
 	VkClearValue clear_values[2];
 	memset(&clear_values, 0, sizeof(clear_values));
-	clear_values[0].color.uint32[0] = 0;
-	clear_values[0].color.uint32[1] = 0;
-	clear_values[0].color.uint32[2] = 0;
-	clear_values[0].color.uint32[3] = 1;
+	clear_values[0].color.float32[0] = 0.0f;
+	clear_values[0].color.float32[1] = 0.0f;
+	clear_values[0].color.float32[2] = 0.0f;
+	clear_values[0].color.float32[3] = 1.0f;
 	clear_values[1].depthStencil.depth = 1.0f;
 	clear_values[1].depthStencil.stencil = 0;
